Upgrade to Spark 1.6.0
